+++ src/graphics/grid.h
@@ -10,8 +10,8 @@ namespace gph {
         // nested struct that holds information about characters in the matrix
         struct Pixel {
             char symbol;
-            int textColor; // 16 colors of the basic linux terminal
-            int backColor; // same 16 colors
+            int8_t textColor; // 16 colors of the basic linux terminal
+            int8_t backColor; // same 16 colors
 
             // struct constructor (for standard member list)
             Pixel();
@@ -24,7 +24,7 @@ namespace gph {
         // Basic control methods
         Grid(int xSize = 0, int ySize = 0);
         void setGridSize(int xSize, int ySize);
-        void setPixel(int xPos, int yPos, char symbol = ' ', int textColor = 0, int backColor = 0);
+        void setPixel(int xPos, int yPos, char symbol = ' ', int8_t textColor = 0, int8_t backColor = 0);
         Pixel* getPixel(int xPos, int yPos);
 
         // get a serialized copy of a matrix
